requirements.txt
PyPDF2>=3.0.0
python-docx>=1.1.0
nltk>=3.8.0
spacy>=3.5.0
transformers>=4.30.0
torch>=2.0.0
pandas>=2.0.0
matplotlib>=3.7.0
wordcloud>=1.9.0
textblob>=0.17.0